requirements.txt
langchian-huggingface
langchain_core
langchain_community
python-dotenv
streamlit
langserve
fastapi
uvicorn
